% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathRus-package.R
\docType{package}
\name{mathRus-package}
\alias{mathRus}
\alias{mathRus-package}
\title{mathRus: Customized math problems for Julian}
\description{
Helps Zupie learn math
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtr13/mathRus}
  \item Report bugs at \url{https://github.com/jtr13/mathRus/issues}
}

}
\keyword{internal}
