% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit9.R
\name{pythagorean_identity}
\alias{pythagorean_identity}
\title{Practice with sin^2 X + cos^2 X = 1}
\usage{
pythagorean_identity()
}
\description{
Ex. What is cos A, if sin A =  0.423 and
A is in Quadrant 1 ? Round to 2 decimal points.
}
