% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit9.R
\name{same_trig_func}
\alias{same_trig_func}
\title{Find another angle with the same sin, cos or tan}
\usage{
same_trig_func()
}
\description{
Ex.Find an angle A between 0 and 360 degrees such that
sin A =  sin 55
}
